requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
cloudinary==1.32.0
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
drf-extensions==0.7.1
idna==3.4
importlib-metadata==6.0.0
Markdown==3.4.1
Pillow==9.4.0
psycopg2-binary==2.9.6
Pygments==2.14.0
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
zipp==3.15.0
